tests: uid() à NULL, eq() ambigu, soldes d'Ardoise faux
Trois défauts de la campagne, dont un qui masquait tout le reste.

`auth.uid()` ne résolvait personne. La stack v0.8.2 le définit par
`nullif(current_setting('request.jwt.claim.sub', true), '')::uuid` : il ne lit
que le GUC scalaire, pas le JSON `request.jwt.claims`. La fixture ne posait que
le JSON, donc uid() valait NULL, aucune politique ne reconnaissait personne et
chaque comptage sous `authenticated` rendait zéro. PostgREST pose les deux
formes à chaque requête — la production n'est pas concernée — mais
testkit.as_user les pose désormais toutes les deux, par symétrie avec lui. Et
0002 vérifie auth.uid() avant sa première assertion : « Alice voit 0 membre »
ne disait rien de la cause, « auth.uid() ne résout pas le sous » la dit.

Les deux surcharges de testkit.eq introduites au tour précédent sont
retirées : elles rendaient `eq(integer, integer, text)` AMBIGU, puisque ni
elles ni la version polymorphe neprimitive mieux que l'autre. Le bon correctif
est ailleurs — la version polymorphe gère très bien les paires de même type, et
seules trois comparaisons `(bigint, integer)` échouaient. Elles sont castées.

L'Ardoise : les attentes du test ne tenaient pas compte de l'arithmétique de
son propre scénario. Bob payait le cinéma (30 €) et en gérait 30 € de parts
(20 + 10) : son solde est exactement zéro, pas -20 €, et le total des soldes
imposait -40 € pour Lina. La fonction de calcul était juste ; c'est le test
qui était faux. Le payeur du cinéma passe à Lina pour qu'aucun solde ne soit
nul — simplify_household_debts écarte les soldes sous 0,005, et la
compensation n'aurait produit qu'un seul transfert, sans exercer le glissement
des deux plus grands soldes que la fonction annonce.

+++ supabase/tests/0001_schema_contract.sql
@@ -277,7 +277,7 @@ insert into testkit.fx (key, user_id)
 values ('bob', testkit.auth_user('bob@example.fr', 'Bob Martin'));
 
 select testkit.eq(
-  (select count(*) from public.profiles p join testkit.fx on testkit.fx.user_id = p.id where testkit.fx.key = 'bob'), 1,
+  (select count(*) from public.profiles p join testkit.fx on testkit.fx.user_id = p.id where testkit.fx.key = 'bob'), 1::bigint,
   'le trigger sur auth.users doit créer le profil');
 select testkit.eq(
   (select p.display_name from public.profiles p join testkit.fx on testkit.fx.user_id = p.id where testkit.fx.key = 'bob'),
@@ -318,7 +318,7 @@ do $$
 begin
   perform testkit.eq(
     (select count(*) from storage.buckets where id in ('household-media', 'household-avatars') and public = false),
-    2,
+    2::bigint,
     'les buckets de médias doivent exister et être privés'
   );
   perform testkit.ok(
